Отладка программы в IDEA[#214983]

+++ chapter_001/src/main/java/ru/job4j/condition/SqArea.java
@@ -2,7 +2,7 @@ package ru.job4j.condition;
 
 public class SqArea {
 
-    public static double square(int p, int k) {
+    public static int square(int p, int k) {
 
         int h = p / (2 *(k + 1));
         int L = h * k;
